RIS-15 Cover order confirmation with e2e-tests (#11)

+++ apps/inventory-microservice/src/app/common/entities/product-stock.entity.ts
@@ -17,7 +17,7 @@ export class ProductStock {
   @Column()
   public quantity: number;
 
-  @Column({ nullable: true, type: 'int' })
+  @Column({ nullable: true, type: 'bigint' })
   public orderProductId: number | null;
 
   @CreateDateColumn()
